move npts back to gluonts.model
