+++ src/gluonts/model/npts/_predictor.py
